refactor(db-config): change connection pool settings to int32 type
- Updated the database connection pool configuration to use int32 for MaxConns and MinConns across various files, ensuring consistency in type handling.
- Modified environment variable parsing functions to return int32 values, enhancing type safety and preventing potential overflow issues.
- Adjusted related tests to reflect the new int32 type for connection pool settings, improving test accuracy and reliability.

This refactor enhances the robustness of database connection configurations and aligns with best practices for type usage.

+++ pre-processor/app/driver/db_connection.go
@@ -66,8 +66,8 @@ func InitPreProcessorDB(ctx context.Context) (*pgxpool.Pool, error) {
 		return nil, fmt.Errorf("failed to parse pre-processor DB config: %w", err)
 	}
 
-	config.MaxConns = int32(dbConfig.MaxConns)
-	config.MinConns = int32(dbConfig.MinConns)
+	config.MaxConns = dbConfig.MaxConns
+	config.MinConns = dbConfig.MinConns
 	config.MaxConnLifetime = 1 * time.Hour
 	config.MaxConnIdleTime = 30 * time.Minute
 	config.ConnConfig.Tracer = &QueryTracer{}
